+++ classes/Aliases.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Neat\Service;
+
+interface Aliases
+{
+    /**
+     * Get aliases
+     *
+     * @return array
+     */
+    public function aliases(): array;
+}